AUTO-345 Jenkins Master Project Plugin
Broke out the Latest Build list into it's own jelly, and made a jelly for each
row in that list.  This way we can just refresh broken rows instead of refreshing the whole list all the time, and we can avoid blind refreshes.

Next change will have the actually refresh logic in it.