AUTO-345 Jenkins Master Project Plugin
Added global abort.  You can now click the abort button on the master build
and all sub-builds will get killed as well.

Also, the Builders need to be one per build and not one per project,
and the CompletionService needs to be one per build.  Previously,
if an abort was called, the watcher threads would not be removed
from the completion service, as the completion service code is in the main
thread which was killed by abort.  This led to subsequent builds taking
the wrong futures from the completeion service.