Remove the forward() from the rebuild AJAX call.
We don't know why the forward exists. It raises a 404 error in the JS console,
so the URL it's trying to forward to apparently doesn't exist/make sense. And
we don't know what work there is to do at the point of the forward() call
anyway, since rebuilding has already been kicked off.