Fix endless blinking orb in MasterResult.isBuilding().
The method used to return true if there are no scheduled sub-jobs. I've removed
that check. The original intent of the check was to start off the blinking orb
right away, even when sub-jobs haven't been scheduled yet, as a corrective to
new logic that based isBuilding() on the status of sub-jobs. With that new
logic, when you rebuilt a sub-job, the orb would start blinking again.

But the emptiness check introduced a bug: if sub-jobs failed to schedule, the
orb on the parent (master) job would blink forever, since the set of sub-jobs
stayed permanently empty. So remove the emptiness check.

But doesn't this reintroduce the issue where the orb won't start blinking until
sub-jobs start running? No, because the call to super.isBuilding() that I
previously added to MasterBuild.isBuilding() already solves that issue. I only
now realize that the emptiness check in MasterResult.isBuilding() is
superfluous (and buggy).